backend/requirement.txt
astropy
numpy
fastapi
uvicorn
python-multipart
pillow
tensorflow
tifffile
scipy
